for extented for unsigned long

+++ utils.h
@@ -10,4 +10,4 @@ void die(const char *fmt);
 
 #define in() {note("Calling => %s",__func__);}
 #define out() {note("Returing <= %s",__func__);}
-#define FOR(number) for (int i = 0; i < number; i++)
+#define FOR(number) for (unsigned long i = 0; i < number; i++)
